1.LongTensor does not take device parameter 2.--local-rank is not recognied arguments which is now deprecated,use   run instread of launch 3.user bfloat16 for float16/half is not supported on Gaudi

+++ generate_samples.py
@@ -257,7 +257,7 @@ def generate_samples(model, tokenizer, args, device):
     if not os.path.exists(output_path):
         os.makedirs(output_path)
     output_path = os.path.join(output_path, f"sample-{datetime.now().strftime('%m-%d-%H-%M')}.txt")
-    with torch.no_grad(), open(output_path, "w") as output, torch.autocast("hpu", dtype=torch.float16, enabled=True):
+    with torch.no_grad(), open(output_path, "w") as output, torch.autocast("hpu", dtype=torch.bfloat16, enabled=True):
         while True:
             torch.distributed.barrier(group=mpu.get_model_parallel_group())
